+++ src/main/java/com/woowacourse/zzinbros/user/web/support/UserArgumentResolver.java
@@ -1,6 +1,5 @@
-package com.woowacourse.zzinbros.user.config;
+package com.woowacourse.zzinbros.user.web.support;
 
-import com.woowacourse.zzinbros.user.domain.UserSession;
 import com.woowacourse.zzinbros.user.exception.UserNotLoggedInException;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
